+++ src/main/java/dao/interfaceDao/MagazzinoDao.java
@@ -12,5 +12,5 @@ public interface MagazzinoDao {
 	public void delete(int idFarmacia, int idProdotto) throws SQLException;
 	public List<MagazzinoDettaglioDTO> getProdottiFarmacia(int idfarmacia) throws SQLException;
 	public int getCountProdotti(int idfarmacia) throws SQLException;
-	public int getPrezzo(int idFarmacia, int idProdotto) throws SQLException;
+	public double getPrezzo(int idFarmacia, int idProdotto) throws SQLException;
 }
